Comments and minor edits
