+++ javascript/webapp/public/schema/codebook/v1.0.0.json
@@ -27,6 +27,7 @@
             "type": "object"
         },
         "Constraints": {
+            "description": "Validation constraints for a field.",
             "properties": {
                 "enum": {
                     "description": "Strict list of allowed values in the column.",
@@ -84,6 +85,7 @@
             "type": "string"
         },
         "Field": {
+            "description": "Contains the full schema definition and metadata for a data field (usually a table column).\nThis includes the required data type, various data nature and rendering properties, potential\nvalidation rules, and mappings from a data dictionary.",
             "properties": {
                 "constraints": {
                     "$ref": "#/definitions/Constraints",
@@ -151,6 +153,7 @@
             "type": "object"
         },
         "FieldMetadata": {
+            "description": "Holds core metadata/stats for a data field.",
             "properties": {
                 "bins": {
                     "items": {
@@ -275,7 +278,7 @@
             "type": "string"
         }
     },
-    "description": "This contains all of the field-level details for interpreting a data table,\nincluding data types, mapping, and metadata.\nTable schemas are standalone resources that can be linked from a Dataset or embedded.\nNote also that with persisted metadata and field examples, a table can often be visualized\nand described to the user without actually loading the source file.",
+    "description": "This contains all of the field-level details for interpreting a data table,\nincluding data types, mapping, and metadata.\nTable schemas are standalone resources that can be linked from a Dataset or embedded.\nNote also that with persisted metadata and field examples, a table can often be visualized\nand described to the user without actually loading the source file.\nresource profile: 'codebook'",
     "properties": {
         "$schema": {
             "type": "string"
