+++ javascript/webapp/public/schema/workflow/v1.5.0.json
@@ -864,7 +864,7 @@
             "type": "string"
         }
     },
-    "description": "The root wrangling workflow specification.\nprofile: 'workflow'",
+    "description": "The root wrangling workflow specification.\nresource profile: 'workflow'",
     "properties": {
         "$schema": {
             "type": "string"
