+++ schema/releases/datapackage/v1.0.0.json
@@ -56,7 +56,7 @@
 			"type": "object"
 		}
 	},
-	"description": "Defines a Data Package, which is a collection of data resources\nsuch as files and schemas.\nLoosely based on the Frictionless spec, but modified where needed to meet our needs.\nhttps://specs.frictionlessdata.io/data-package/",
+	"description": "Defines a Data Package, which is a collection of data resources such as files and schemas.\nLoosely based on the Frictionless spec, but modified where needed to meet our needs.\nhttps://specs.frictionlessdata.io/data-package/",
 	"properties": {
 		"$schema": {
 			"type": "string"
