+++ schema/releases/workflow/v1.5.0.json
@@ -830,7 +830,7 @@
 			"type": "string"
 		}
 	},
-	"description": "The root wrangling workflow specification.\nprofile: 'workflow'",
+	"description": "The root wrangling workflow specification.\nresource profile: 'workflow'",
 	"properties": {
 		"$schema": {
 			"type": "string"
